+++ TestSupport/EfSchemeCompare/Internal/Stage2Comparer.cs
@@ -12,7 +12,7 @@ namespace TestSupport.EfSchemeCompare.Internal
     internal class Stage2Comparer
     {
         private readonly DatabaseModel _databaseModel;
-        private readonly StringComparer _stringComparer;
+        private readonly StringComparer _caseComparer;
 
         private bool _hasErrors;
 
@@ -24,7 +24,7 @@ namespace TestSupport.EfSchemeCompare.Internal
         {
             _databaseModel = databaseModel;
             _ignoreList = config?.LogsToIgnore ?? new List<CompareLog>();
-            _stringComparer = config?.CaseComparer ?? StringComparer.CurrentCulture;
+            _caseComparer = config?.CaseComparer ?? StringComparer.CurrentCulture;
         }
 
         public bool CompareLogsToDatabase(IReadOnlyList<CompareLog> firstStageLogs)
@@ -40,12 +40,12 @@ namespace TestSupport.EfSchemeCompare.Internal
 
         private void LookForUnusedTables(IReadOnlyList<CompareLog> firstStageLogs, CompareLog log)
         {
-            var logger = new CompareLogger(CompareType.Table, null, _logs, _ignoreList, () => _hasErrors = true);
+            var logger = new CompareLogger(CompareType.Table, null, log.SubLogs, _ignoreList, () => _hasErrors = true);
             var databaseTableNames = _databaseModel.Tables.Select(x => x.FormSchemaTable(_databaseModel.DefaultSchema));
             var allEntityTableNames = firstStageLogs.SelectMany(p => p.SubLogs)
                 .Where(x => x.State == CompareState.Ok && x.Type == CompareType.Entity)
                 .Select(p => p.Expected).OrderBy(p => p).Distinct().ToList();
-            var tablesNotUsed = databaseTableNames.Where(p => !allEntityTableNames.Contains(p));
+            var tablesNotUsed = databaseTableNames.Where(p => !allEntityTableNames.Contains(p, _caseComparer));
 
             foreach (var tableName in tablesNotUsed)
             {
@@ -56,14 +56,14 @@ namespace TestSupport.EfSchemeCompare.Internal
         private void LookForUnusedColumns(IReadOnlyList<CompareLog> firstStageLogs, CompareLog log)
         {
             var logger = new CompareLogger(CompareType.Column, null, _logs, _ignoreList, () => _hasErrors = true);
-            var tableDict = _databaseModel.Tables.ToDictionary(x => x.FormSchemaTable(_databaseModel.DefaultSchema));
+            var tableDict = _databaseModel.Tables.ToDictionary(x => x.FormSchemaTable(_databaseModel.DefaultSchema), _caseComparer);
             //because of table splitting and TPH we need to groups properties by table name to correctly say what columns are missed
             var entityColsGrouped = firstStageLogs.SelectMany(p => p.SubLogs)
                 .Where(x => x.State == CompareState.Ok && x.Type == CompareType.Entity)
                 .GroupBy(x => x.Expected, y => y.SubLogs
                     .Where(x => x.State == CompareState.Ok && x.Type == CompareType.Property)
                     .Select(p => p.Expected));
-            var entityColsByTableDict = entityColsGrouped.ToDictionary(x => x.Key, y => y.SelectMany(x => x.ToList()));
+            var entityColsByTableDict = entityColsGrouped.ToDictionary(x => x.Key, y => y.SelectMany(x => x.ToList()), _caseComparer);
 
             foreach (var entityLog in firstStageLogs.SelectMany(p => p.SubLogs)
                 .Where(x => x.State == CompareState.Ok && x.Type == CompareType.Entity))
@@ -71,7 +71,7 @@ namespace TestSupport.EfSchemeCompare.Internal
                 if (tableDict.ContainsKey(entityLog.Expected))
                 {
                     var dbColNames = tableDict[entityLog.Expected].Columns.Select(x => x.Name);
-                    var colsNotUsed = dbColNames.Where(p => !entityColsByTableDict[entityLog.Expected].Contains(p));
+                    var colsNotUsed = dbColNames.Where(p => !entityColsByTableDict[entityLog.Expected].Contains(p, _caseComparer));
                     foreach (var colName in colsNotUsed)
                     {
                         logger.ExtraInDatabase(colName, CompareAttributes.ColumnName, entityLog.Expected);
@@ -83,7 +83,7 @@ namespace TestSupport.EfSchemeCompare.Internal
         private void LookForUnusedIndexes(IReadOnlyList<CompareLog> firstStageLogs, CompareLog log)
         {
             var logger = new CompareLogger(CompareType.Index, null, _logs, _ignoreList, () => _hasErrors = true);
-            var tableDict = _databaseModel.Tables.ToDictionary(x => x.FormSchemaTable(_databaseModel.DefaultSchema));
+            var tableDict = _databaseModel.Tables.ToDictionary(x => x.FormSchemaTable(_databaseModel.DefaultSchema), _caseComparer);
             foreach (var entityLog in firstStageLogs.SelectMany(p => p.SubLogs)
                 .Where(x => x.State == CompareState.Ok && x.Type == CompareType.Entity))
             {
@@ -93,7 +93,7 @@ namespace TestSupport.EfSchemeCompare.Internal
                     var allEfIndexNames = entityLog.SubLogs
                         .Where(x => x.State == CompareState.Ok && x.Type == CompareType.Index)
                         .Select(p => p.Expected).OrderBy(p => p).Distinct().ToList();
-                    var indexesNotUsed = indexCol.Where(p => !allEfIndexNames.Contains(p));
+                    var indexesNotUsed = indexCol.Where(p => !allEfIndexNames.Contains(p, _caseComparer));
                     foreach (var indexName in indexesNotUsed)
                     {
                         logger.ExtraInDatabase(indexName, CompareAttributes.IndexConstraintName, entityLog.Expected);